Add optional lock indicator for output statusline
When g:vorax_output_lock_indicator is 1 (default is 0) then show the
status of TM and TX locks on statusline. So you known whether you need
to perform still rollback or commit.
When status show in lowercase then no locks for this type exists for own
session. But when uppercase then exists locks of this type.

You must have permission to select on oracle system view v$lock.